Add common base for pct and pick strategies
